examples/django-next-notes/backend/pyproject.toml
[project]
name = "django-next-notes-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "django>=4.0",
  "djangorestframework>=3.14",
  "django-cors-headers>=4.0",
  "psycopg2-binary>=2.9",
]
